View components, Server config, MongoDB
Implement a Passport session to store user's sessions to determine factors.
Implement a login and sign up process using passport, bycrypt for authentication and also Google Oauth as a secondary authentication.
Implemented a pagination page to divide articles received per page
Implemented a search login for all user's queries and parameters
Implemented a settings page to save and display user's preferred categories.
Implemented a NavBar to display to login menu
Implemented the blog page to display general news when user is not signed in